requirements.txt
faiss-cpu==1.7.4
fastapi==0.95.1
langchain==0.0.161
openai==0.27.6
pypdf2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
tiktoken==0.3.3
uvicorn==0.22.0